Refactor: Move mailbox related classed to mailbox package. Make constructor public.
