+++ core/src/main/java/bisq/core/alert/PrivateNotificationMessage.java
@@ -17,8 +17,8 @@
 
 package bisq.core.alert;
 
-import bisq.network.p2p.MailboxMessage;
 import bisq.network.p2p.NodeAddress;
+import bisq.network.p2p.mailbox.MailboxMessage;
 
 import bisq.common.app.Version;
 import bisq.common.proto.network.NetworkEnvelope;
