+++ core/src/main/java/bisq/core/trade/messages/PayoutTxPublishedMessage.java
@@ -19,8 +19,8 @@ package bisq.core.trade.messages;
 
 import bisq.core.account.sign.SignedWitness;
 
-import bisq.network.p2p.MailboxMessage;
 import bisq.network.p2p.NodeAddress;
+import bisq.network.p2p.mailbox.MailboxMessage;
 
 import bisq.common.app.Version;
 import bisq.common.proto.network.NetworkEnvelope;
