+++ core/src/main/java/bisq/core/trade/protocol/TradeProtocol.java
@@ -28,9 +28,9 @@ import bisq.network.p2p.AckMessage;
 import bisq.network.p2p.AckMessageSourceType;
 import bisq.network.p2p.DecryptedDirectMessageListener;
 import bisq.network.p2p.DecryptedMessageWithPubKey;
-import bisq.network.p2p.MailboxMessage;
 import bisq.network.p2p.NodeAddress;
 import bisq.network.p2p.SendMailboxMessageListener;
+import bisq.network.p2p.mailbox.MailboxMessage;
 import bisq.network.p2p.messaging.DecryptedMailboxListener;
 
 import bisq.common.Timer;
