+++ p2p/src/main/java/bisq/network/p2p/P2PService.java
@@ -20,6 +20,9 @@ package bisq.network.p2p;
 import bisq.network.Socks5ProxyProvider;
 import bisq.network.crypto.EncryptionService;
 import bisq.network.p2p.mailbox.IgnoredMailboxService;
+import bisq.network.p2p.mailbox.MailboxItem;
+import bisq.network.p2p.mailbox.MailboxMessage;
+import bisq.network.p2p.mailbox.MailboxMessageList;
 import bisq.network.p2p.messaging.DecryptedMailboxListener;
 import bisq.network.p2p.network.CloseConnectionReason;
 import bisq.network.p2p.network.Connection;
