+++ p2p/src/test/java/bisq/network/p2p/mocks/MockMailboxPayload.java
@@ -17,8 +17,8 @@
 
 package bisq.network.p2p.mocks;
 
-import bisq.network.p2p.MailboxMessage;
 import bisq.network.p2p.NodeAddress;
+import bisq.network.p2p.mailbox.MailboxMessage;
 import bisq.network.p2p.storage.payload.ExpirablePayload;
 
 import bisq.common.app.Version;
